pyproject.toml
[project]
name = "product-hunt"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.21.0",
    "rich>=14.2.0",
    "tinydb>=4.8.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
